+++ Toggly.FeatureManagement.Python/toggly-django/src/toggly_django/__init__.py
@@ -10,7 +10,7 @@ from toggly_django.decorators import feature_flag_required, feature_gate_require
 from toggly_django.middleware import TogglyMiddleware
 from toggly_django.utils import get_client, get_context_from_request
 
-__version__ = "0.1.0"
+__version__ = "0.3.0"
 __all__ = [
     "TogglyDjangoConfig",
     "TogglyMiddleware",