+++ Toggly.FeatureManagement.Python/toggly-fastapi/src/toggly_fastapi/__init__.py
@@ -34,7 +34,7 @@ from toggly_fastapi.middleware import (
     get_toggly_client,
 )
 
-__version__ = "0.1.0"
+__version__ = "0.3.0"
 __all__ = [
     "ClientDep",
     "ContextDep",